created res policy during item installation, tests
